+++ meetup2apricot/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Joel Shprentz"""
 __email__ = "jshprentz@his.com"
-__version__ = "1.1.0"
+__version__ = "1.2.0"