+++ HISTORY.rst
@@ -9,7 +9,6 @@ Future History
 * Detect photo file type by inspection; Meetup always claims jpeg.
 * Accept multiple accounting codes from event title.
 * Tag featured events.
-* Restrict registrations for events labeled members only.
 * Update tags and member-only status of existing events.
 * Limit processed events by last start date.
 
@@ -17,8 +16,10 @@ Future History
 Recent History
 ~~~~~~~~~~~~~~
 
-.. Next Release
-.. ------------------
+Next Release
+------------------
+
+* Restrict registrations for events labeled members only.
 
 1.1.0 (2020-11-17)
 ------------------